Split util.js file and move all the utilities to distinct folder.
