+++ src/components/sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 export const createSortedTemplate = () => {
   return (
